Rename modules (#90)
