+++ src/lib.rs
@@ -1,9 +1,9 @@
 mod camera;
 #[cfg(feature = "dev")]
-mod dev;
+mod dev_tools;
 mod game;
 mod screen;
-mod util;
+mod ui_tools;
 
 use bevy::{asset::AssetMetaCheck, prelude::*};
 
@@ -35,10 +35,15 @@ impl Plugin for AppPlugin {
         );
 
         // Add other plugins.
-        app.add_plugins((game::plugin, screen::plugin, util::plugin, camera::plugin));
+        app.add_plugins((
+            game::plugin,
+            screen::plugin,
+            ui_tools::plugin,
+            camera::plugin,
+        ));
 
         // Enable dev tools for dev builds.
         #[cfg(feature = "dev")]
-        app.add_plugins(dev::plugin);
+        app.add_plugins(dev_tools::plugin);
     }
 }
